requirements.txt
apache-airflow==2.10.5
pandas==2.2.3
requests==2.32.3
apache-airflow-providers-http==5.3.1
apache-airflow-providers-sqlite==4.0.0
